+++ src/precompile.jl
@@ -159,6 +159,6 @@ let
     end
 
     if Base.generating_output()
-        pkg_precompile()
+        # pkg_precompile()
     end
 end