+++ test/test_packages/WithSources/LocalPkg/src/LocalPkg.jl
@@ -1,0 +1,5 @@
+module LocalPkg
+
+greet() = print("Hello World!")
+
+end # module LocalPkg