+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Pkg v1.12 Release Notes
+=======================
+
+- It is now possible to specify "sources" for packages in a `[sources]` section in Project.toml.
+  This can be used to add non-registered normal or test dependencies.
+
 Pkg v1.11 Release Notes
 =======================